Final commit for satisfying requirement
These are all required files to run the system. The server performs all the operations at its end and maintains logs. And client file is user interactive file to which shows the operations and images before and after process completed. Proto Buffer is file of format and structure of system.